Update text and rename MonitorDefinitionCard directory
